+++ lib/appium_lib/android/element/button.rb
@@ -14,6 +14,7 @@ module Appium
         index = value
         raise "#{index} is not a valid index. Must be >= 1" if index <= 0
 
+        # 1 indexed
         return find_element :uiautomator, _button_visible_selectors(index: index)
       end
 
@@ -74,7 +75,6 @@ module Appium
       elements.first
     end
 
-    # @private
     def _button_visible_selectors(opts = {})
       button_index       = opts.fetch :button_index, false
       image_button_index = opts.fetch :image_button_index, false
@@ -88,14 +88,12 @@ module Appium
       end
     end
 
-    # @private
     def _button_exact_string(value)
       button       = string_visible_exact Button, value
       image_button = string_visible_exact ImageButton, value
       button + image_button
     end
 
-    # @private
     def _button_contains_string(value)
       button       = string_visible_contains Button, value
       image_button = string_visible_contains ImageButton, value