+++ lib/appium_lib/android/espresso/element.rb
@@ -1,11 +1,11 @@
-require_relative 'espresso/helper'
-require_relative 'espresso/element'
-require_relative 'espresso/bridge'
+require_relative 'element/generic'
+require_relative 'element/button'
 
 module Appium
   module Android
     module Espresso
-      # parent
+      module Element
+      end # module Element
     end # module Espresso
   end # module Android
 end # module Appium